# docs/migration/contracts/api.yaml
...
openapi: 3.1.0
info:
  title: Silicon Extend API
  version: 2.0.0-draft
  description: |
    Silicon Extend lets a Silicon use a Carbon's personal devices. This API is used by the `extend` CLI
    (through the `silicon-extend-client` crate), the configuration website, and the Extend apps on devices.

    Conventions
    - Every JSON body is an envelope: `{"type": "<kind>", "data": {...}}`.
    - Errors: `{"type": "error", "data": {"code", "message", "hint", "docs_url", "request_id", "details"}}`.
      `code` is stable and machine-readable; `message` says what happened and why; `hint` says what to do next.
    - Timestamps are RFC 3339 UTC with milliseconds. Durations carry their unit in the field name.
    - Every response carries `X-Request-ID` and `Silicon-Extend-API-Version`.
    - Lists are paged with `limit` and `cursor`; the response gives `next_cursor`, null on the last page. A page that
      has a `next_cursor` is always full; only the last page can be short.
    - Retry a POST after an uncertain result with the same `Idempotency-Key` and the identical body. The same key with
      a different body is `409 conflict`.
    - API majors (TECHNICAL.md section 10). Every `/api/v{n}/` request is checked before anything else: a major this
      service doesn't serve is `400 api_version_unsupported`; a `Silicon-Extend-API-Version` pin naming another major
      than the path is `400 api_version_mismatch`; a sunset major is `410 api_version_sunset`, whose hint says how to
...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/contracts:
    get:
      tags: [System]
      operationId: contracts
      summary: Compatibility matrix and deprecation state
      description: |
        Built from live state, not a fixed list. Each major's lifecycle (`current` → `deprecated` → `sunset`) is kept in
        `extend_global.api_versions`, shared by every instance: configuration deprecates a major
        (`EXTEND_DEPRECATED_API_VERSIONS`, applied when an instance starts; the newest served major can't be
        deprecated), and a deprecated major is sunset automatically after 7 consecutive days with zero requests
        (requests are counted per major and UTC day; an instance re-checks every 5 minutes). Sunset is final. The client
        crate and CLI ranges come from `EXTEND_API_V{n}_CLIENT_CRATE` and `EXTEND_API_V{n}_CLI` (default
        `>={n}.0.0, <{n+1}.0.0`), and `device_app_min` from `EXTEND_DEVICE_APP_MIN_VERSION`. Retired majors stay listed
        after their routes are gone. A request to this route counts as a request today for the major it came in on.
        `extend version` reads this to say whether the CLI is current, deprecated or sunset.
      security: []
      responses:
        '200':
          description: Every API major, its state, and the client, CLI and app versions that work with it
          headers:
            Deprecation: {$ref: '#/components/headers/Deprecation'}
            Sunset: {$ref: '#/components/headers/Sunset'}
          content:
            application/json:
              schema:
                type: object
                required: [type, data]
                properties:
                  type: {const: contracts}
                  data:
                    type: object
                    required: [service_version, current, supported, deprecated, sunset_rule, versions]
                    properties:
                      service_version: {type: string, example: 1.0.0}
                      current: {type: integer, example: 1, description: The newest major a client can agree on}
                      supported: {type: array, items: {type: integer}, example: [1], description: Majors served and not sunset}
                      deprecated: {type: array, items: {type: integer}, example: []}
                      sunset_rule: {type: string, example: 'Sunset after 7 consecutive days with zero requests'}
                      versions:
                        type: array
                        description: Every major with a lifecycle record, oldest first, including retired ones
                        items:
                          type: object
                          required: [api_version, state, deprecated_at, sunset_at, sunset_earliest_at, last_request_on, sunset_rule, compatible]
                          properties:
                            api_version: {type: integer}
                            state: {type: string, enum: [current, deprecated, sunset]}
                            deprecated_at: {type: [string, 'null'], format: date-time}
                            sunset_at: {type: [string, 'null'], format: date-time, description: When the sunset rule was met}
                            sunset_earliest_at:
                              type: [string, 'null']
                              format: date-time
                              description: 'Deprecated majors only: the soonest it can be sunset if no request arrives (the Sunset header)'
                            last_request_on: {type: [string, 'null'], format: date, description: The last UTC day with a request}
                            sunset_rule: {type: string, example: 'Sunset after 7 consecutive days with zero requests'}
                            compatible:
                              type: object
                              required: [client_crate, cli, device_app_min]
                              properties:
                                client_crate: {type: string, example: '>=1.0.0, <2.0.0'}
                                cli: {type: string, example: '>=1.0.0, <2.0.0'}
                                device_app_min: {type: string, example: 1.0.0}
              example:
                type: contracts
                data:
                  service_version: 1.0.0
                  current: 1
                  supported: [1]
                  deprecated: []
                  sunset_rule: Sunset after 7 consecutive days with zero requests
                  versions:
                    - api_version: 1
                      state: current
                      deprecated_at: null
                      sunset_at: null
                      sunset_earliest_at: null
                      last_request_on: '2026-09-27'
                      sunset_rule: Sunset after 7 consecutive days with zero requests
                      compatible: {client_crate: '>=1.0.0, <2.0.0', cli: '>=1.0.0, <2.0.0', device_app_min: 1.0.0}
        '410': {description: 'api_version_sunset: the path''s major is retired', $ref: '#/components/responses/Error'}
  /api/v2/auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: Sign out of Extend
      description: |
        Signs this sign-in out of Extend. Silicon Accounts revokes the refresh token sent in the body (the CLI and the
        website hold it), or, without one, the sign-in of the access token used. Then:
        - a Silicon's running sessions end at once (`silicon_logged_out`), its open wake requests are withdrawn, and the
          proofs Extend held for it are revoked;
        - a Carbon's sign-out ends the sessions of the Silicons they gave access to, through their own pairs only
          (`access_removed`), never another Carbon's.
        Pairs and grants stay. If Silicon Accounts can't be reached, nothing is revoked or ended and the error says so.
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              required: [type, data]
              properties:
                type: {const: logout}
                data:
                  type: object
                  additionalProperties: false
                  properties:
                    refresh_token: {type: string, writeOnly: true, minLength: 1, description: 'The Silicon Accounts refresh token (sar_…) the client holds'}
      responses:
        '204': {description: Signed out}
        '401': {$ref: '#/components/responses/Error'}
        '503': {description: 'Silicon Accounts could not be reached; nothing was revoked or ended', $ref: '#/components/responses/Error'}
  /api/v2/me:
    get:
      tags: [Auth]
      operationId: me
      summary: Who is signed in (what `extend login status --json` prints)
      description: |
        Who the access token belongs to: the account's permanent Silicon Accounts `uuid`, its current public id (which can
        change; Extend keys nothing on it), its kind and, for a Silicon, its custodian.
      responses:
        '200':
          description: The account the access token belongs to
          content:
            application/json:
              schema:
                type: object
                required: [type, data]
                properties:
                  type: {const: me}
                  data: {$ref: '#/components/schemas/AccountMe'}
        '401': {$ref: '#/components/responses/Error'}

  # ───────────────────────────── Enrollment (unpaired app) ─────────────────────────────
  /api/v1/enrollments:
    post:
      tags: [Enrollment]
      operationId: createEnrollment
      summary: An unpaired Extend app asks for a pairing code
      description: |
        No sign-in. An Extend app that isn't paired yet asks for a 6-character pairing code and an enrollment secret, then
        holds `/api/v1/enrollments/{enrollment_id}/connect` open to get new codes and, once a Carbon claims one, its device
        credential. An app older than `EXTEND_DEVICE_APP_MIN_VERSION` is `426 upgrade_required`. At most 10 enrollments per
        address every 10 minutes. The device wire is unchanged in Extend 4: apps installed before it keep working.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, data]
              properties:
                type: {const: enrollment}
                data:
                  type: object
                  additionalProperties: false
                  required: [os, app_version]
                  properties:
                    os: {$ref: '#/components/schemas/DeviceOs'}
                    os_version: {type: string, maxLength: 64, example: '15'}
                    model: {type: string, maxLength: 128, example: Pixel 9}
                    app_version: {type: string, maxLength: 32, example: 1.1.0}
                    engine_version: {type: string, maxLength: 32, description: 'The device engine''s version on a computer (1.1). 1.0 apps send it as agent_device_version, which is still read.'}
                    agent_device_version: {type: string, maxLength: 32, deprecated: true, description: 'The 1.0 name of engine_version'}
      responses:
        '201':
          description: Enrollment with its first code
          content:
            application/json:
              schema:
                type: object
                required: [type, data]
                properties:
                  type: {const: enrollment}
                  data:
                    type: object
                    required: [enrollment_id, enrollment_secret, pairing_code, code_expires_at, rotates_every_s]
                    properties:
                      enrollment_id: {type: string, format: uuid}
                      enrollment_secret: {type: string, pattern: '^ees_[A-Za-z0-9_-]{43}$', description: Returned once}
                      pairing_code: {$ref: '#/components/schemas/PairingCode'}
                      code_expires_at: {type: string, format: date-time}
                      rotates_every_s: {type: integer, const: 300}
              example:
                type: enrollment
                data:
                  enrollment_id: 01926f3a-5c1e-7b2d-9a40-3e5f8c7d1b22
                  enrollment_secret: ees_Q2hhbmdlLW1lLXRoaXMtaXMtYS1zYW1wbGUtc2VjcmV
                  pairing_code: 4F9C2A
                  code_expires_at: '2026-09-26T10:05:00.000Z'
                  rotates_every_s: 300
        '401': {description: 'testing_secret_invalid: the request still selects a test environment (Extend 4 has none)', $ref: '#/components/responses/Error'}
        '422': {description: 'invalid_input, for example an OS that pairs through a host (ios, tvos, samsung_tv…)', $ref: '#/components/responses/Error'}
        '426': {$ref: '#/components/responses/Error'}
        '429': {$ref: '#/components/responses/Error'}
  /api/v1/enrollments/{enrollment_id}:
    parameters:
      - {$ref: '#/components/parameters/EnrollmentId'}
    get:
      tags: [Enrollment]
      operationId: getEnrollment
      summary: Current code, or the device credential once paired
      description: |
        The enrollment secret as `Authorization: Extend-Enrollment <secret>`. While unclaimed: the current code. Once a
        Carbon claimed it: the device credential, exactly once; the enrollment is then gone and later reads are 404.
      security:
        - enrollmentSecret: []
      responses:
        '200':
          description: Enrollment state
          content:
            application/json:
              schema:
                type: object
                required: [type, data]
                properties:
                  type: {const: enrollment}
                  data: {$ref: '#/components/schemas/EnrollmentState'}
        '401': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
    delete:
      tags: [Enrollment]
      operationId: discardEnrollment
      summary: The app stops waiting to be paired
      security:
        - enrollmentSecret: []
      responses:
        '204': {description: Discarded; its code no longer works}
        '401': {$ref: '#/components/responses/Error'}
  /api/v1/enrollments/{enrollment_id}/connect:
    parameters:
      - {$ref: '#/components/parameters/EnrollmentId'}
    get:
      tags: [Enrollment]
      operationId: enrollmentSocket
      summary: WebSocket that pushes code rotations and the pairing result
      description: |
        Upgrade to WebSocket. Frames are JSON `{"type", "data"}`:
        `code` `{pairing_code, code_expires_at}` every 300 s; `paired` `{device_id, device_credential, environment}`
        once, then the server closes with 1000. Heartbeat as on the device socket.
      security:
        - enrollmentSecret: []
      responses:
        '101': {description: Switching protocols}
        '401': {$ref: '#/components/responses/Error'}

  # ───────────────────────────── Pairing (Carbon) ─────────────────────────────
  /api/v2/pairings:
    post:
      tags: [Pairing]
      operationId: claimPairingCode
      summary: A Carbon enters a device's pairing code
      description: |
        Carbons only. Claims a pairing code the Extend app shows: the device becomes the caller's pair of it. The code works
        once. A wrong code never reveals whether it exists; 5 wrong codes in 15 minutes are `429 rate_limited`.
        - `silicon_ids` gives those Silicons access at once: any active Silicon, by its current `si:` id or its Silicon
          Accounts uuid. An id nobody has, a Carbon's id, or a Silicon its custodian hasn't accepted yet is `422
          invalid_input`, and nothing is paired.
        - A device is private to the Carbon who paired it and the Silicons they give access to: `visibility` other than
          `personal` is `422 invalid_input`.
        - The Carbon is enrolled to receive Extend's notifications through Ting (when it is on and they aren't yet).
        "Pair with another Carbon" makes a second pair of the same physical device for another Carbon; each pair has its own
        id, name, access list, lifetime and activity log.
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, data]
              properties:
                type: {const: pairing}
                data:
                  type: object
                  additionalProperties: false
                  required: [pairing_code, name]
                  properties:
                    pairing_code: {$ref: '#/components/schemas/PairingCodeInput'}
                    name: {$ref: '#/components/schemas/DeviceName'}
                    visibility: {$ref: '#/components/schemas/Visibility'}
                    pair_ttl_days: {$ref: '#/components/schemas/PairTtlDays'}
                    silicon_ids:
                      type: array
                      maxItems: 100
                      uniqueItems: true
                      items: {$ref: '#/components/schemas/SiliconId'}
                      description: Optional; give access now. Can be skipped and done later.
            example: {type: pairing, data: {pairing_code: 4f9c2a, name: Saket's Pixel, silicon_ids: ['si:chef']}}
      responses:
        '201':
          description: Paired. The device is in `setup` until its own setup steps finish.
          headers:
            ETag: {schema: {type: string}}
          content:
            application/json:
              schema:
                type: object
                required: [type, data]
                properties:
                  type: {const: device}
                  data: {$ref: '#/components/schemas/Device'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {description: 'pairing_code_invalid: wrong, expired or used', $ref: '#/components/responses/Error'}
        '409': {description: 'test_device_limit; conflict (1.1: you already paired this device, named by your own id for it)', $ref: '#/components/responses/Error'}
        '422': {description: 'invalid_input: an unknown or unaccepted Silicon in silicon_ids, or visibility other than personal', $ref: '#/components/responses/Error'}
        '429': {description: 'rate_limited: too many wrong codes, or the test environment is busy adding other devices', $ref: '#/components/responses/Error'}
  /api/v2/devices/{device_id}/attachments:
    parameters:
      - {$ref: '#/components/parameters/DeviceId'}
    post:
      tags: [Pairing]
      operationId: attachDevice
      summary: Pair an iPhone, iPad, Apple TV or smart TV through a paired computer
      description: |
        Owner only: adds a device the caller's computer carries (an Apple TV, an iPhone, a smart TV) through the caller's
        own pair of that computer. The host must be a computer that runs the app itself and is online; a carried device
        can't carry others. `visibility` other than `personal` is `422`. The host is told to set it up (`attach`).
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, data]
              properties:
                type: {const: attachment}
                data:
                  type: object
                  additionalProperties: false
                  required: [os, name]
                  properties:
                    os: {type: string, enum: [ios, ipados, tvos, samsung_tv, lg_tv]}
                    name: {$ref: '#/components/schemas/DeviceName'}
                    visibility: {$ref: '#/components/schemas/Visibility'}
                    pair_ttl_days: {$ref: '#/components/schemas/PairTtlDays'}
      responses:
        '201':
          description: Attached device, in setup
          content:
            application/json:
              schema:
                type: object
                required: [type, data]
                properties:
                  type: {const: device}
                  data: {$ref: '#/components/schemas/Device'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {description: test_device_limit, $ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '429': {description: 'rate_limited: the test environment is busy adding other devices', $ref: '#/components/responses/Error'}
        '503': {description: device_offline (the host), $ref: '#/components/responses/Error'}
  /api/v2/devices/{device_id}/setup:
    parameters:
      - {$ref: '#/components/parameters/DeviceId'}
    get:
      tags: [Pairing]
      operationId: getSetup
      summary: The device's setup steps and where each one stands
      description: |
        Owner only (the Carbon who paired this pair). Also answers for a removed device the caller paired.
      responses:
        '200':
          description: Setup state
          content:
            application/json:
              schema:
                type: object
                required: [type, data]
                properties:
                  type: {const: setup}
                  data: {$ref: '#/components/schemas/Setup'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
...
  /api/v2/devices/{device_id}/access:
    parameters:
      - {$ref: '#/components/parameters/DeviceId'}
    get:
      tags: [Access]
      operationId: listAccess
      summary: Silicons that can use the device
      description: |
        Owner only. Every grant on this pair, each with the Silicon's current id and permanent `silicon_uuid`, who gave it,
        and `wake_muted`. A removed device the caller paired answers with no grants.
      responses:
        '200':
          description: Access grants
          content:
            application/json:
              schema:
                type: object
                required: [type, data]
                properties:
                  type: {const: access}
                  data:
                    type: object
                    required: [items]
                    properties:
                      items: {type: array, items: {$ref: '#/components/schemas/AccessGrant'}}
        '403': {$ref: '#/components/responses/Error'}
...
  /api/v2/contracts:
    get:
      tags: [System]
      operationId: contractsV2
      summary: Compatibility matrix and deprecation state (the same answer as /api/v1/contracts)
      security: []
      responses:
        '200': {description: 'Every API major, its state, and the client, CLI and app versions that work with it'}
        '410': {description: 'api_version_sunset: the path''s major is retired', $ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A Silicon Accounts access token for the app `extend`: an EdDSA-signed JWT whose `aud` is `extend` and whose
        `iss` is the Silicon Accounts this server trusts (`ACCOUNTS_URL`), verified with its published keys. Carbons get
        one with the Silicon Accounts device flow (the CLI) or the authorization-code flow (the website's server),
        Silicons by exchanging a short-lived token from `silicon-accounts login --app extend -q`. Tokens issued before
        a sign-out, removed access or deleted account are refused.
    enrollmentSecret:
      type: http
      scheme: bearer
      bearerFormat: 'Enrollment secret (ees_). Sent as "Authorization: Extend-Enrollment <secret>".'
    deviceCredential:
      type: http
      scheme: bearer
      bearerFormat: 'Device credential (edc_). Sent as "Authorization: Extend-Device <credential>".'
...